refactor(packages): delete dead packages + unify duplicate utilities
P1: Remove packages/db tombstone (README+.gitkeep only, zero importers)
P2: Remove packages/desktop-client tombstone (README+.gitkeep only, zero importers)
P3: Fold buildDesktopDataDir into packages/desktop-core, delete packages/desktop-shell (zero importers confirmed)
P4: Delete duplicate packages/interfaces/src/core/credit.interface.ts — ICredit/ICreditEntry/ICreditsBreakdown already exported via billing/credits.interface.ts in main barrel
P5: Canonicalize getErrorMessage — delete app-local copy in apps/server/api/src/helpers/utils/error, redirect all callers to @libs/utils/error; remove inline copy from packages/agent, re-export from @genfeedai/utils/error/error-handler.util
P6: Wire packages/cli to @genfeedai/errors — add workspace dep, remove inlined BaseCliError/ApiError/formatError, keep CLI-specific error classes extending BaseCliError

All turbo builds green (42/42 tasks excluding Next.js apps); type-check 43/43 tasks; biome check clean; CLI 118 tests, agent 281 tests, utils 271 tests all passing.
