+++ apps/server/api/src/endpoints/webhooks/klingai/webhooks.kling.service.ts
@@ -1,8 +1,8 @@
 import { MetadataService } from '@api/collections/metadata/services/metadata.service';
-import { getErrorMessage } from '@api/helpers/utils/error/get-error-message.util';
 import { KlingAIWebhookPayload } from '@libs/interfaces/webhook-payload.interface';
 import { LoggerService } from '@libs/logger/logger.service';
 import { CallerUtil } from '@libs/utils/caller/caller.util';
+import { getErrorMessage } from '@libs/utils/error/get-error-message.util';
 import { Injectable } from '@nestjs/common';
 
 type MediaUrls = {
