+++ apps/server/api/src/endpoints/webhooks/services/post-processing-orchestrator.service.ts
@@ -2,11 +2,11 @@ import { EvaluationsService } from '@api/collections/evaluations/services/evalua
 import { type IngredientDocument } from '@api/collections/ingredients/schemas/ingredient.schema';
 import { OrganizationSettingsService } from '@api/collections/organization-settings/services/organization-settings.service';
 import { ConfigService } from '@api/config/config.service';
-import { getErrorMessage } from '@api/helpers/utils/error/get-error-message.util';
 import { UserExtractionUtil } from '@api/helpers/utils/user-extraction/user-extraction.util';
 import { BotGatewayService } from '@api/services/bot-gateway/bot-gateway.service';
 import { EvaluationType, IngredientCategory } from '@genfeedai/enums';
 import { LoggerService } from '@libs/logger/logger.service';
+import { getErrorMessage } from '@libs/utils/error/get-error-message.util';
 import { Injectable, Optional } from '@nestjs/common';
 
 @Injectable()
