+++ packages/cli/package.json
@@ -5,6 +5,7 @@
     "gf": "dist/index.js"
   },
   "dependencies": {
+    "@genfeedai/errors": "workspace:*",
     "@genfeedai/serializers": "workspace:*",
     "@genfeedai/tools": "workspace:*",
     "@inquirer/prompts": "8.5.2",
