--- packages/db/README.md
@@ -1,5 +1,0 @@
-# DB
-
-Reserved database package directory.
-
-This directory is currently a placeholder. Do not move database logic here casually; formalize the package first if shared DB code is needed.
