--- packages/desktop-client/README.md
@@ -1,5 +1,0 @@
-# Desktop Client
-
-Reserved desktop client package directory.
-
-This directory is currently a placeholder for future shared desktop client code.
