+++ packages/desktop-core/src/index.ts
@@ -1,6 +1,17 @@
+import fs from 'node:fs';
 import path from 'node:path';
 
 const METADATA_DIR = '.genfeed';
+
+export function buildDesktopDataDir(appName: string): string {
+  const baseDir = process.env.GENFEED_DESKTOP_DATA_DIR
+    ? path.resolve(process.env.GENFEED_DESKTOP_DATA_DIR)
+    : path.join(process.cwd(), '.data');
+
+  const appDir = path.join(baseDir, appName);
+  fs.mkdirSync(appDir, { recursive: true });
+  return appDir;
+}
 
 export function buildWorkspaceMetadataDir(root: string): string {
   return path.join(root, METADATA_DIR);
