--- packages/desktop-shell/README.md
@@ -1,5 +1,0 @@
-# Desktop Shell
-
-Desktop shell entrypoint package.
-
-This package is the shared shell-facing layer for desktop runtime composition and wrappers.
