--- packages/desktop-shell/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@genfeedai/desktop-shell",
-  "version": "0.1.1",
-  "private": true,
-  "dependencies": {
-    "@genfeedai/desktop-prisma": "workspace:*"
-  },
-  "main": "./src/index.ts",
-  "type": "module",
-  "types": "./src/index.ts"
-}
